+++ src/constants.ts
@@ -1,20 +1,19 @@
 import { h } from "./utils";
 
-export const DEFAULT_ORIGIN = "https://www.youtube.com";
-
+export const DEFAULT_ORIGIN = h`68747470733a2f2f7777772e796f75747562652e636f6d`;
 export const DEFAULT_HEADERS = {
   "Accept-Language": "en",
   "User-Agent":
     "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/90.0.4430.93 Safari/537.36",
 };
-
 export const DC = {
   clientName: "WEB",
   clientVersion: "2.20210813.00.00",
 };
-
 export const DAK = h`41497a615379414f5f464a32536c7155385134535445484c4743696c775f59395f313171635738"`;
-
 export const EP_GLCR = h`2f796f7574756265692f76312f6c6976655f636861742f6765745f6c6976655f636861745f7265706c6179`;
-
 export const EP_GLC = h`2f796f7574756265692f76312f6c6976655f636861742f6765745f6c6976655f63686174`;
+export const XO = h`582d4f726967696e`;
+export const XGAU = h`582d476f6f672d4175746855736572`;
+export const XGPID = h`582d476f6f672d506167654964`;
+export const SASH = h`5341504953494448415348`;