+++ src/index.ts
@@ -27,7 +27,11 @@ export {
 export { YTReloadContinuation } from "./yt/context";
 
 export interface MasterchatOptions {
+  /** you can grab Credentials using `extra/credential-fetcher` */
   credentials?: Credentials | string;
+
+  sessionId?: string;
+
   isLive?: boolean;
 }
 
@@ -51,12 +55,13 @@ export class Masterchat {
   constructor(
     videoId: string,
     channelId: string,
-    { credentials, isLive = true }: MasterchatOptions = {}
+    { isLive, credentials, sessionId }: MasterchatOptions = {}
   ) {
     this.videoId = videoId;
     this.channelId = channelId;
     this.isLive = isLive;
     this.apiKey = DAK;
+    this.sessionId = sessionId;
 
     if (typeof credentials === "string") {
       credentials = JSON.parse(